+++ apps/starter/CHANGELOG.md
@@ -1,5 +1,16 @@
 # stoker
 
+## 0.4.52
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/web-client@0.5.23
+    - @stoker-platform/utils@0.5.21
+    - @stoker-platform/web-app@0.5.43
+    - @stoker-platform/cli@0.5.38
+    - @stoker-platform/node-client@0.5.27
+
 ## 0.4.51
 
 ### Patch Changes