+++ packages/node-client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stoker-platform/node-client
 
+## 0.5.27
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/utils@0.5.21
+
 ## 0.5.26
 
 ### Patch Changes